refactor: use AsyncAPI 3.1 schemas from
