+++ packages/workspace-store/src/schemas.ts
@@ -1,4 +1,3 @@
 // biome-ignore lint/performance/noBarrelFile: This is a barrel file that re-exports workspace schemas
-export { AsyncApiDocument } from './schemas/asyncapi/asyncapi-document'
 export { isAsyncApiDocument, isOpenApiDocument } from './schemas/type-guards'
-export { type Workspace, type WorkspaceDocument, getWorkspaceDocumentSchema } from './schemas/workspace'
+export type { Workspace, WorkspaceDocument } from './schemas/workspace'
