+++ packages/workspace-store/src/schemas/type-guards.ts
@@ -1,4 +1,6 @@
-import type { AsyncApiDocument } from './asyncapi/asyncapi-document'
+import { isObject } from '@scalar/helpers/object/is-object'
+import type { AsyncApiDocument } from '@scalar/types/asyncapi/3.1'
+
 import type { OpenApiDocument } from './v3.1/strict/openapi-document'
 
 /**
@@ -9,10 +11,7 @@ import type { OpenApiDocument } from './v3.1/strict/openapi-document'
  * or broader contexts that mix documents with the workspace itself).
  */
 export const isOpenApiDocument = (value: unknown): value is OpenApiDocument =>
-  typeof value === 'object' &&
-  value !== null &&
-  'openapi' in value &&
-  typeof (value as { openapi: unknown }).openapi === 'string'
+  isObject(value) && 'openapi' in value && typeof value.openapi === 'string'
 
 /**
  * Narrow a value to an AsyncAPI document.
@@ -20,10 +19,7 @@ export const isOpenApiDocument = (value: unknown): value is OpenApiDocument =>
  * Discriminated by the required `asyncapi` string field on AsyncAPI documents.
  */
 export const isAsyncApiDocument = (value: unknown): value is AsyncApiDocument =>
-  typeof value === 'object' &&
-  value !== null &&
-  'asyncapi' in value &&
-  typeof (value as { asyncapi: unknown }).asyncapi === 'string'
+  isObject(value) && 'asyncapi' in value && typeof value.asyncapi === 'string'
 
 /**
  * Identify the document type of a value.
