+++ packages/workspace-store/src/schemas/workspace.ts
@@ -1,6 +1,6 @@
 import { Type } from '@scalar/typebox'
+import type { AsyncApiDocument } from '@scalar/types/asyncapi/3.1'
 import { AVAILABLE_CLIENTS, type AvailableClients } from '@scalar/types/snippetz'
-import { type Schema, union } from '@scalar/validation'
 
 import { compose } from '@/schemas/compose'
 import { extensions } from '@/schemas/extensions'
@@ -16,27 +16,13 @@ import { type XScalarCookies, xScalarCookiesSchema } from '@/schemas/extensions/
 import { type XScalarOrder, XScalarOrderSchema } from '@/schemas/extensions/general/x-scalar-order'
 import { type XScalarActiveProxy, XScalarActiveProxySchema } from '@/schemas/extensions/workspace/x-scalar-active-proxy'
 import { type XScalarTabs, XScalarTabsSchema } from '@/schemas/extensions/workspace/x-scalar-tabs'
-import { generateSchema as generateOpenApiSchema } from '@/schemas/v3.1/openapi'
 
-import { AsyncApiDocument } from './asyncapi/asyncapi-document'
 import type { OpenAPIExtensions, OpenApiDocument } from './v3.1/strict/openapi-document'
 
 export type WorkspaceDocumentMeta = Omit<
   OpenAPIExtensions,
   'x-original-oas-version' | 'x-scalar-original-source-url' | 'x-scalar-original-document-hash'
 >
-
-/**
- * A document held by the workspace — a discriminated union of OpenAPI and AsyncAPI.
- * Narrow with the `isOpenApiDocument` / `isAsyncApiDocument` guards in `@/schemas/type-guards`
- * before touching spec-specific fields.
- *
- * The OpenAPI half is built via {@link generateOpenApiSchema}, which takes a `maybeRef`
- * callback so callers can plug in their own `$ref` handling (the in-memory store passes
- * `recursiveRef`; consumers parsing raw input can pass an identity function).
- */
-export const getWorkspaceDocumentSchema = (maybeRef: (inner: Schema) => Schema): Schema =>
-  union([generateOpenApiSchema(maybeRef), AsyncApiDocument], { typeName: 'WorkspaceDocument' })
 
 export type WorkspaceDocument = OpenApiDocument | AsyncApiDocument
